pyproject.toml
[project]
name = "book-learn"
version = "0.1.0"
description = "im making the worlds most complicated book app(as in follow all the best practices and write clean code)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.7.0",
    "asyncpg==0.30.0",
    "certifi==2024.12.14",
    "click==8.1.8",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "fastapi-cli==0.0.7",
    "fastapi==0.115.6",
    "greenlet==3.1.1",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "jinja2==3.1.5",
    "loadenv==0.1.1",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "psycopg2==2.9.10",
    "psycopg2-binary==2.9.10",
    "pydantic==2.10.4",
    "pydantic-settings==2.7.0",
    "pydantic-core==2.27.2",
    "pygments==2.18.0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "rich==13.9.4",
    "rich-toolkit==0.12.0",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.36",
    "sqlmodel==0.0.22",
    "starlette==0.41.3",
    "typer==0.15.1",
    "typing-extensions==4.12.2",
    "uvicorn==0.34.0",
    "uvloop==0.21.0",
    "watchfiles==1.0.3",
    "websockets==14.1",
    "alembic>=1.14.0",
    "passlib>=1.7.4",
    "bcrypt>=4.2.1",
    "pyjwt[crypto]>=2.10.1",
    "redis>=5.2.1",
]
